requirements.txt
pytest==7.1.3
Appium-Python-Client==2.6.1
selenium
allure-pytest
pytest-xdist
requests